Moving _blip_rotate_scale to __init__ to avoid code repetition.
